docs: add documentation

+++ estructuras/RTreeNode.hpp
@@ -3,10 +3,13 @@
 
 #include "Rect.hpp"
 
+/**
+ * @brief Estructura que representa un nodo interno del RTree.
+*/
 struct RTreeNode {
   Rect MBR;
-  unsigned int firstChildIndex;
-  unsigned int lastChildIndex;
+  long long firstChildIndex;
+  long long lastChildIndex;
 };
 
 #endif
